requirements.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
albumentations==1.0.3
antlr4-python3-runtime==4.8
anyio==3.3.2
appdirs==1.4.4
argon2-cffi==21.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.7.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==21.4b2
bleach==1.5.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.5.18.1
cffi==1.14.6
chamfer-distance @ git+https://github.com/otaheri/chamfer_distance@d2b524309db114d0f7ce18be6c01b3802cde9791
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
cryptography==36.0.2
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
dglgo==0.0.2
docutils==0.19
dpcpp-cpp-rt==2022.1.0
entrypoints==0.3
fastjsonschema==2.15.3
Flask==2.0.3
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.0
future==0.18.2
fvcore==0.1.5.post20211023
gast==0.4.0
glumpy==1.2.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.3.0
html5lib==0.9999999
hydra-core==1.1.1
idna==3.3
imageio==2.9.0
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==4.11.3
importlib-resources==5.2.2
intel-cmplr-lib-rt==2022.1.0
intel-cmplr-lic-rt==2022.1.0
intel-opencl-rt==2022.1.0
intel-openmp==2022.1.0
iopath==0.1.8
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.5
jupyter-core==4.8.1
jupyter-packaging==0.10.6
jupyter-server==1.11.0
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.1.1
kaolin==0.11.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
littleutils==0.2.2
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mkl==2022.1.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==1.5.0
mmcv-full==1.5.3
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
ninja==1.10.2.3
nose==1.3.7
notebook==6.4.4
notebook-shim==0.1.0
numpy
# numpy==1.21.5
numpydoc==1.4.0
oauthlib==3.2.0
ogb==1.3.4
olefile==0.46
omegaconf==2.1.1
open3d==0.13.0
opencv-python==4.5.2.52
opencv-python-headless==4.5.2.52
opt-einsum==3.3.0
outdated==0.2.1
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
plyfile==0.7.4
portalocker==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.4
psutil==5.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pydot==1.4.2
Pygments==2.10.0
PyJWT==2.4.0
PyOpenGL==3.1.6
pyOpenSSL==22.0.0
pyparsing==2.4.7
pypng==0.20220715.0
pyquaternion==0.9.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@276c9a8acbfa715f5802f26ec9f4141bde26ecb4
pytz==2021.1
pyu2f==0.1.5
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
rdkit-pypi==2022.3.5
regex==2021.8.28
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.18.3
scikit-learn==1.0
scipy
# scipy==1.7.3
Send2Trash==1.8.0
Shapely==1.7.1
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tabulate==0.8.9
tbb==2021.6.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.1
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.8.30
tinycss2==1.1.1
toml==0.10.2
tomlkit==0.7.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.0
triangle==20220202
typer==0.6.1
typing_extensions==4.2.0
urllib3==1.26.9
usd-core==22.5.post1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.1.2
widgetsnbextension==3.5.1
wrapt==1.14.1
yacs==0.1.8
yapf==0.31.0
yarl==1.7.2
zipp==3.8.0
